cctv-service/src/utils/auth.py
from jose import jwt, JWTError
from datetime import datetime, timedelta
from typing import Optional
import os

JWT_SECRET = os.getenv("JWT_SECRET", "your-super-secret-jwt-key-change-this-in-production")
JWT_ALGORITHM = "HS256"

def create_token(user_id: str, expires_delta: Optional[timedelta] = None) -> str:
    """Create JWT token"""
    to_encode = {"userId": user_id}
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(hours=24)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, JWT_SECRET, algorithm=JWT_ALGORITHM)
    return encoded_jwt

def verify_token(token: str) -> Optional[str]:
    """Verify JWT token and return user_id"""
    try:
        payload = jwt.decode(token, JWT_SECRET, algorithms=[JWT_ALGORITHM])
        user_id: str = payload.get("userId")
        if user_id is None:
            return None
        return user_id
    except JWTError:
        return None

def decode_token(token: str) -> Optional[dict]:
    """Decode JWT token and return full payload"""
    try:
        payload = jwt.decode(token, JWT_SECRET, algorithms=[JWT_ALGORITHM])
        return payload
    except JWTError:
        return None
