cctv-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
redis==5.0.1
httpx==0.25.0
websockets==12.0
aiofiles==23.2.1
pydantic==2.4.2
python-jose[cryptography]==3.3.0
asyncio-mqtt==0.16.2
aiortc==1.6.0
av==11.0.0
psutil==5.9.6
prometheus-client==0.19.0
prisma==0.11.0
python-dotenv==1.0.0
